requirements.txt
streamlit
streamlit_option_menu
pandas 
numpy
pillow
streamlit_extras.dataframe_explorer